Moving all dofile to require
