+++ MUSHclient/worlds/plugins/DBUMMO_Chat_Window.xml
@@ -4,7 +4,7 @@
 <muclient>
 <plugin
    name="DBUMMO_Chat_Window"
-   author="Fiendish"
+   author="Braska"
    id="b555825a4a5700c35fa80780"
    language="Lua"
    purpose="Move chats to a miniwindow"
@@ -19,7 +19,7 @@
 
 Thanks to Fiendish of Aardwolf who provided the base and starting point for all of this.
 Original files sourced from: https://github.com/fiendish/aardwolfclientpackage
-My process: http://www.gammon.com.au/forum/?id=14718
+
 Thank you!!
 
 My Github : https://github.com/Xinefus-Braska
@@ -312,14 +312,13 @@ USAGE:
 
 <script>
 <![CDATA[
---dofile(GetPluginInfo(GetPluginID(), 20) .. "register_z_on_create.lua")
 
-dofile(GetPluginInfo(GetPluginID(), 20) .. "mw_theme_base.lua")
+require "register_z_on_create"
+require "mw_theme_base"
 require "movewindow"
-dofile(GetPluginInfo(GetPluginID(), 20) .. "text_rect.lua")
-dofile(GetPluginInfo(GetPluginID(), 20) .. "scrollbar.lua")
-
-dofile(GetPluginInfo(GetPluginID(), 20) .. "DBUMMO_colors.lua")
+require "text_rect"
+require "scrollbar"
+require "DBUMMO_colors"
 
 local gap = 1
 MAX_LINES = 10000 -- how many lines to store in scrollback
