+++ MUSHclient/worlds/plugins/DBUMMO_Theme_Controller.xml
@@ -14,7 +14,7 @@
 
 Thanks to Fiendish of Aardwolf who provided the base and starting point for all of this.
 Original files sourced from: https://github.com/fiendish/aardwolfclientpackage
-My process: http://www.gammon.com.au/forum/?id=14718
+
 Thank you!!
 
 My Github : https://github.com/Xinefus-Braska
@@ -37,7 +37,7 @@ My Github : https://github.com/Xinefus-Braska
 <script>
 <![CDATA[
 
-dofile(GetPluginInfo(GetPluginID(), 20) .. "mw_theme_base.lua")
+require "mw_theme_base"
 
 theme_file = GetVariable("theme_file") or Theme.theme_file
 SetVariable("theme_file", theme_file)
