--- MUSHclient/worlds/plugins/DBUMMO_register_z_on_create.lua
@@ -1,5 +1,0 @@
-WindowCreate = function(w, ...)
-  ret = world.WindowCreate(w, ...)
-  CallPlugin("462b665ecb569efbf261422f", "registerMiniwindow", w)
-  return ret
-end
